+++ lib/grape_oas/exporter/oas2/response.rb
@@ -34,7 +34,7 @@ module GrapeOAS
         def build_schema_or_ref(schema)
           if schema.respond_to?(:canonical_name) && schema.canonical_name
             @ref_tracker << schema.canonical_name if @ref_tracker
-            ref_name = schema.canonical_name.gsub("::", "_")
+            ref_name = GrapeOAS.schema_ref_name.call(schema.canonical_name)
             { "$ref" => "#/definitions/#{ref_name}" }
           else
             Schema.new(schema, @ref_tracker, nullable_strategy: @nullable_strategy).build